+++ R/altimetry-files.R
@@ -28,7 +28,7 @@ altimetry_currents_polar_files <- function(hemisphere = "south") {
   ff <- split(files, grepl("polar_v", files$fullname))
   files <- ff[[1]] %>% dplyr::rename(ufullname = .data$fullname) %>% dplyr::inner_join(ff[[2]] %>%
                                                                                    dplyr::rename(vfullname = .data$fullname), "date") %>%
-    dplyr::select(.data$date, .data$ufullname, .data$vfullname)
+    dplyr::select("date", "ufullname", "vfullname")
 
   dplyr::arrange(dplyr::distinct(files, .data$date, .keep_all = TRUE), .data$date)
 }