+++ R/ncep2-files.R
@@ -22,7 +22,7 @@ ncep2_uwnd_6hr_files <- function() {
 
   if (nrow(files) < 1)
     stop("no files found")
-  files <- dplyr::transmute(files, date = ISOdate(as.integer(stringr::str_extract(basename(fullname), "[0-9]{4}")), 1, 1, 0, 0, 0, tz = "UTC"),
+  files <- dplyr::transmute(files, date = ISOdate(as.integer(stringr::str_extract(basename(.data$fullname), "[0-9]{4}")), 1, 1, 0, 0, 0, tz = "UTC"),
                          fullname = .data$fullname, root = .data$root)
   dplyr::arrange(dplyr::distinct(files, .data$date, .keep_all = TRUE), .data$date)
 
@@ -35,7 +35,7 @@ ncep2_vwnd_6hr_files <- function() {
 
   if (nrow(files) < 1)
     stop("no files found")
-  files <- dplyr::transmute(files, date = ISOdate(as.integer(stringr::str_extract(basename(fullname), "[0-9]{4}")), 1, 1, tz  = "UTC"),
+  files <- dplyr::transmute(files, date = ISOdate(as.integer(stringr::str_extract(basename(.data$fullname), "[0-9]{4}")), 1, 1, tz  = "UTC"),
                          fullname = .data$fullname, root = .data$root)
   dplyr::arrange(dplyr::distinct(files, .data$date, .keep_all = TRUE), .data$date)